+++ yellowbox/extras/webserver/util.py
@@ -2,6 +2,9 @@ from __future__ import annotations
 
 from contextlib import contextmanager
 from logging import Filter, getLogger
+from typing import Awaitable, Callable, Iterable, TypeVar, Union
+
+from starlette.responses import Response
 
 
 def reason_is_ne(field: str, expected, got) -> str:
@@ -43,3 +46,30 @@ def mute_uvicorn_log():
     logger.addFilter(filter_)
     yield
     logger.removeFilter(filter_)
+
+
+T = TypeVar('T')
+
+
+def iter_side_effects(side_effects: Iterable[Union[Callable[..., Awaitable[T]], T]]) -> Callable[..., Awaitable[T]]:
+    """
+    Args:
+        side_effects: An iterable of side effects.
+
+    Returns:
+        A side effect that varies from call to call. On the first call, delegating to the first side effect on
+         side_effects, on the second to the second, and so on.
+
+    Notes:
+        This function respects starlette responses in that if it encounters one, it will simply return it instead of
+         delegating to it.
+    """
+    tor = iter(side_effects)
+
+    async def ret(*args, **kwargs):
+        next_side_effect = next(tor)
+        if isinstance(next_side_effect, Response):  # responses are async callable smh
+            return next_side_effect
+        return await next_side_effect(*args, **kwargs)
+
+    return ret